Resolved all imports/moved back to ThemePreferences.ts
