+++ components/hub/ColorPreferences.ts
@@ -1,3 +1,12 @@
+import { HumanColorName, HColor } from '@Hi/Colors';
+import ClickButton from '@Hi/Components/ClickButton';
+import HStack from '@Hi/Components/HStack';
+import IonIcon from '@Hi/Components/IonIcon';
+import Spacer from '@Hi/Components/Spacer';
+import TextView from '@Hi/Components/TextView';
+import VStack from '@Hi/Components/VStack';
+import highlightColorSelected from '@Triggers/preferences-triggers';
+
 /**
  * Creates a circle icon with a specified color. The button binds to a
  * handler which assigns the browser's color theme.
@@ -6,7 +15,7 @@
  * @class HighlightColorButton
  * @extends {ClickButton}
  */
-export default class HighlightColorButton extends ClickButton {
+export class HighlightColorButton extends ClickButton {
     /**
      * Creates an instance of HighlightColorButton.
      * @param {HumanColorName} color The color of the highlight button and the
@@ -31,7 +40,7 @@ export default class HighlightColorButton extends ClickButton {
  * @class HighlightColorPreferences
  * @extends {VStack}
  */
-export default class HighlightColorPreferences extends VStack {
+export class HighlightColorPreferences extends VStack {
     /**
      * Creates an instance of HighlightColorPreferences.
      *
@@ -64,68 +73,6 @@ export default class HighlightColorPreferences extends VStack {
                         'yellow',
                     ] as HumanColorName[]
                 ).map(color => new HighlightColorButton(color)),
-                new Spacer(),
-            ).width('100%'),
-        );
-        this.width('100%').padding();
-    }
-}
-
-/**
- * Allows the user to select either a dark theme or light theme. This will
- * be rendered as a part of the main preferences page in the hub window.
- *
- * @export
- * @class ThemePreferences
- * @extends {VStack}
- */
-export default class ThemePreferences extends VStack {
-    constructor() {
-        super(
-            new HStack(
-                new TextView('Theme').font('md').bold().margin({ bottom: 10 }),
-                new Spacer(),
-            ).width('100%'),
-
-            new HStack(
-                new VStack(
-                    new ClickButton(
-                        Resources.getImageView('LightThemeThumb.png').rounded(),
-                    )
-                        .padding(0)
-                        .border({
-                            size: 1,
-                            style: 'solid',
-                            color: HColor('gray3'),
-                        })
-                        .rounded()
-                        .whenClicked(() => {
-                            changeTheme('light');
-                            flexarch.pref('theme', 'light');
-                        }),
-                    new TextView('Light Mode').margin({ top: 5 }).font('sm'),
-                ).rounded(),
-
-                new Spacer(),
-
-                new VStack(
-                    new ClickButton(
-                        Resources.getImageView('DarkThemeThumb.png').rounded(),
-                    )
-                        .padding(0)
-                        .border({
-                            size: 1,
-                            style: 'solid',
-                            color: HColor('gray3'),
-                        })
-                        .rounded()
-                        .whenClicked(() => {
-                            changeTheme('dark');
-                            flexarch.pref('theme', 'dark');
-                        }),
-                    new TextView('Dark Mode').margin({ top: 5 }).font('sm'),
-                ).rounded(),
-
                 new Spacer(),
             ).width('100%'),
         );
